Fix all clippy warnings
- Add #[allow(dead_code)] to FrameResult enum (Suspend variant not yet used)
- Add #[allow(clippy::large_enum_variant)] to AST enums
- Remove unnecessary casts in array.rs
- Remove unnecessary .to_string() calls, use .as_ref() instead
- Use .contains() for range checks in number.rs
- Remove needless borrows before format!() calls
- Simplify duplicate if-else branches in try-catch
- Use .iter().flatten() instead of manual if-let loops
- Collapse nested if-let patterns
- Add #[allow(clippy::only_used_in_recursion)] for recursive methods
- Use .unwrap_or() instead of .unwrap_or_else() for simple values
- Convert loop-match to while-let in parser
- Derive Default for JsValue instead of manual impl
- Fix test file keys to match import specifiers

+++ src/interpreter/builtins/array.rs
@@ -1283,7 +1283,7 @@ pub fn array_fill(
                 n.min(length)
             }
         })
-        .unwrap_or(length as i64) as u32;
+        .unwrap_or(length) as u32;
 
     for i in start..end {
         arr_ref.set_property(PropertyKey::Index(i), value.clone());
@@ -1344,7 +1344,7 @@ pub fn array_copy_within(
                 n.min(length)
             }
         })
-        .unwrap_or(length as i64) as u32;
+        .unwrap_or(length) as u32;
 
     let elements: Vec<JsValue> = (start..end)
         .map(|i| {
@@ -1413,7 +1413,7 @@ pub fn array_splice(
     let insert_items: Vec<JsValue> = args.into_iter().skip(2).collect();
     let insert_count = insert_items.len() as u32;
 
-    let new_length = (length as u32 - delete_count + insert_count) as u32;
+    let new_length = length as u32 - delete_count + insert_count;
 
     if insert_count > delete_count {
         let shift = insert_count - delete_count;
